Errors are now useable. Ready to code.
Actually removed a large portion of spaghetti mess. The current mess is workable and not nearly as intrusive.
ls now just needs two print functions, and then all of find probably in a day all on adderall
Except all these classes and finals and FUCK
It's good bruh
Super cash